finished version
This is how the game is supposed to look - x's represent the cards faced down since you do not know what the card is.
I took off display(table) and other things which helped me when coding the game but is not necessary when actually playing.
the while loops that start with:
while i ==1    #AI player 2
while i ==2    #AI player 3
while i ==3    #AI player 4
are all pretty much the same thing. Differences will be shown in screen shot under Technical Solution so you do not have to re-read the same thing 3 times.